+++ examples/visualization.py
@@ -2,8 +2,8 @@
 Visualizing the patterns and flows
 ==================================
 
-:class:`~graphix.visualization.GraphVisualizer` tool offers a wide selection of 
-visualization methods for inspecting the causal structure of the graph associated 
+:class:`~graphix.visualization.GraphVisualizer` tool offers a wide selection of
+visualization methods for inspecting the causal structure of the graph associated
 with the pattern, graph or the (generalized-)flow.
 """