+++ graphix/simulator.py
@@ -4,11 +4,12 @@ Simulates MBQC by executing the pattern.
 
 """
 
+from __future__ import annotations
+
 import warnings
 
 import numpy as np
 
-from graphix.noise_models import NoiseModel
 from graphix.sim.density_matrix import DensityMatrixBackend
 from graphix.sim.statevec import StatevectorBackend
 from graphix.sim.tensornet import TensorNetworkBackend
@@ -46,14 +47,12 @@ class PatternSimulator:
         elif backend == "densitymatrix":
             if noise_model is None:
                 self.noise_model = None
-                # no noise: no need to compute probabilities
                 self.backend = DensityMatrixBackend(pattern, **kwargs)
                 warnings.warn(
                     "Simulating using densitymatrix backend with no noise. To add noise to the simulation, give an object of `graphix.noise_models.Noisemodel` to `noise_model` keyword argument."
                 )
             if noise_model is not None:
                 self.set_noise_model(noise_model)
-                # if noise: have to compute the probabilities
                 self.backend = DensityMatrixBackend(pattern, pr_calc=True, **kwargs)
         elif backend in {"tensornetwork", "mps"} and noise_model is None:
             self.noise_model = None
@@ -86,8 +85,6 @@ class PatternSimulator:
             the output quantum state,
             in the representation depending on the backend used.
         """
-
-        self.backend.add_nodes(self.pattern.input_nodes)
         if self.noise_model is None:
             for cmd in self.pattern:
                 kind = cmd.kind